feat: use dynamic routes for auctions
