+++ components/QRPage.tsx
@@ -4,7 +4,7 @@
 export function QRPage() {
   return (
     <img
-      src="qrLogo.png"
+      src="/qrLogo.png"
       alt="QR Code"
       className="md:flex-1 overflow-hidden w-[220px] h-[220px] md:w-[279px] md:h-[279px]"
     />
